constant changes done for prescription-details and visit-summary
